Updates to French translations, PII, and SSC tools

+++ main_app/views.py
@@ -15,6 +15,7 @@ from azure.core.credentials import AzureKeyCredential
 import re
 from pdfminer.high_level import extract_text
 from presidio_analyzer import AnalyzerEngine
+from presidio_analyzer import recognizer_result
 import fitz  
 
 def banner_image(request):
@@ -26,9 +27,71 @@ def home(request):
     return render(request, 'home.html')
 
 def PII(request):
+    """
+    Renders the home page for the chat application. Gets the URLS for example documents.
+
+    """
+    fs = FileSystemStorage()
+    # Sample 1
+    sample_name = 'demo_employee1.png'
+    filename = os.path.join('cached_outputs', sample_name)
+    example_file_url = fs.url(filename)
+    request.session['sample1_pii_thumbnail_url'] = example_file_url
+    # Strip the .jpg extension and add .pdf  
+    pdf_name = sample_name.rsplit('.', 1)[0] + '.pdf'  
+    request.session['sample1_pii_name_url'] = pdf_name
+
+    # Sample 2
+    sample_name = 'demo_employee2.png'
+    filename = os.path.join('cached_outputs', sample_name)
+    example_file_url = fs.url(filename)
+    request.session['sample2_pii_thumbnail_url'] = example_file_url
+    # Strip the .jpg extension and add .pdf  
+    pdf_name = sample_name.rsplit('.', 1)[0] + '.pdf'  
+    request.session['sample2_pii_name_url'] = pdf_name
+
+    # Sample 3
+    sample_name = 'demo_employee3.png'
+    filename = os.path.join('cached_outputs', sample_name)
+    example_file_url = fs.url(filename)
+    request.session['sample3_pii_thumbnail_url'] = example_file_url
+    # Strip the .jpg extension and add .pdf  
+    pdf_name = sample_name.rsplit('.', 1)[0] + '.pdf'  
+    request.session['sample3_pii_name_url'] = pdf_name
     return render(request, 'PII.html')
 
 def SSC(request):
+    """
+    Renders the home page for the chat application. Gets the URLS for example documents.
+
+    """
+    fs = FileSystemStorage()
+    # Sample 1
+    sample_name = 'demo_employee1.png'
+    filename = os.path.join('cached_outputs', sample_name)
+    example_file_url = fs.url(filename)
+    request.session['sample1_pii_thumbnail_url'] = example_file_url
+    # Strip the .jpg extension and add .pdf  
+    pdf_name = sample_name.rsplit('.', 1)[0] + '.pdf'  
+    request.session['sample1_pii_name_url'] = pdf_name
+
+    # Sample 2
+    sample_name = 'demo_employee2.png'
+    filename = os.path.join('cached_outputs', sample_name)
+    example_file_url = fs.url(filename)
+    request.session['sample2_pii_thumbnail_url'] = example_file_url
+    # Strip the .jpg extension and add .pdf  
+    pdf_name = sample_name.rsplit('.', 1)[0] + '.pdf'  
+    request.session['sample2_pii_name_url'] = pdf_name
+
+    # Sample 3
+    sample_name = 'demo_employee3.png'
+    filename = os.path.join('cached_outputs', sample_name)
+    example_file_url = fs.url(filename)
+    request.session['sample3_pii_thumbnail_url'] = example_file_url
+    # Strip the .jpg extension and add .pdf  
+    pdf_name = sample_name.rsplit('.', 1)[0] + '.pdf'  
+    request.session['sample3_pii_name_url'] = pdf_name
     return render(request, 'SSC.html')
 
      
@@ -76,57 +139,82 @@ def pii_recognition_example(pdf_path):
 # Function to perform the sensitivity analysis
 def analyze_sensitivity(request):
     if request.method == 'POST':
-        uploaded_file = request.FILES['document']
-        fs = FileSystemStorage()
-        filename = fs.save(uploaded_file.name, uploaded_file)
-        document_path = fs.path(filename)
-        analyzer = AnalyzerEngine()
+        if 'example_document' in request.POST:
+            fs = FileSystemStorage()
 
-        # Extract text from PDF
-        score = pii_recognition_example(document_path)
+            document_path = fs.path(os.path.join('cached_outputs', request.POST.get('example_document')))
 
-        # Clean up the file after analysis
-        fs.delete(filename)
+            # Extract text from PDF
+            score = pii_recognition_example(document_path)
 
-        print("Sensitivity Score Calculated: ", score)
+            return JsonResponse({"score": score, "pdfURL": fs.url(os.path.join('cached_outputs', request.POST.get('example_document')))})
+        else:
+            uploaded_file = request.FILES['document']
+            fs = FileSystemStorage()
+            filename = fs.save(uploaded_file.name, uploaded_file)
+            document_path = fs.path(filename)
+            analyzer = AnalyzerEngine()
 
-        return JsonResponse({"score": score})
+            # Extract text from PDF
+            score = pii_recognition_example(document_path)
+
+            # Clean up the file after analysis
+            fs.delete(filename)
+
+            print("Sensitivity Score Calculated: ", score)
+
+            return JsonResponse({"score": score})
 
 
 
 def upload_document2(request):
     if request.method == 'POST':
-        uploaded_file = request.FILES.get('document')
-        if uploaded_file and uploaded_file.name.lower().endswith('.pdf'):
+        if 'example_document' in request.POST:
             fs = FileSystemStorage()
-            
+
             # Split the filename and extension
-            name, extension = os.path.splitext(uploaded_file.name)
+            name, extension = os.path.splitext(request.POST.get('example_document'))
             # Define the new filename with '-redacted' appended
             new_filename = f"{name}-redacted{extension}"
-            
-            # Save the original file
-            filename = fs.save(uploaded_file.name, uploaded_file)
-            uploaded_file_url = fs.url(filename)
-            full_temp_path = os.path.join(fs.location, filename)
-            
-            # Define the output path with the new filename
-            output_pdf_path = os.path.join(fs.location, new_filename)
-            
+
+            full_temp_path = os.path.join(fs.location, os.path.join('cached_outputs', request.POST.get('example_document')))
+            output_pdf_path = os.path.join(fs.location, os.path.join('cached_outputs', new_filename))
+
             # Call your redaction function here
             redact_pdf_based_on_pii(full_temp_path, output_pdf_path)
 
-            # Make sure to delete the original file
-            os.remove(full_temp_path)
-            
-            # Return the path to the redacted PDF with the new filename
-            redacted_pdf_url = fs.url(new_filename)
+            redacted_pdf_url = fs.url(os.path.join('cached_outputs', new_filename))
+            print(redacted_pdf_url)
             return JsonResponse({'redactedPdfUrl': redacted_pdf_url})
         else:
-            return HttpResponseBadRequest("Please upload a valid PDF file.")
+            uploaded_file = request.FILES.get('document')
+            if uploaded_file and uploaded_file.name.lower().endswith('.pdf'):
+                fs = FileSystemStorage()
+                
+                # Split the filename and extension
+                name, extension = os.path.splitext(uploaded_file.name)
+                # Define the new filename with '-redacted' appended
+                new_filename = f"{name}-redacted{extension}"
+                
+                # Save the original file
+                filename = fs.save(uploaded_file.name, uploaded_file)
+                uploaded_file_url = fs.url(filename)
+                full_temp_path = os.path.join(fs.location, filename)
 
- 
- 
+                # Define the output path with the new filename
+                output_pdf_path = os.path.join(fs.location, new_filename)
+
+                # Call your redaction function here
+                redact_pdf_based_on_pii(full_temp_path, output_pdf_path)
+
+                # Make sure to delete the original file
+                os.remove(full_temp_path)
+
+                # Return the path to the redacted PDF with the new filename
+                redacted_pdf_url = fs.url(new_filename)
+                return JsonResponse({'redactedPdfUrl': redacted_pdf_url})
+            else:
+                return HttpResponseBadRequest("Please upload a valid PDF file.")
 
  
  
@@ -142,9 +230,13 @@ def redact_pdf_based_on_pii(pdf_path, output_pdf_path):
 
     # Define PII entities to redact
     entities_to_redact = [
-        'PERSON', 'PHONE_NUMBER', 'EMAIL_ADDRESS', 
-        'CREDIT_CARD', 'IBAN', 'US_BANK_NUMBER', 
-        'DATE_TIME', 'NATIONAL_ID', 'SSN'
+        'PHONE_NUMBER', 'EMAIL_ADDRESS', 'CREDIT_CARD', 'US_DRIVER_LICENSE'
+        # 'CREDIT_CARD', 'IBAN', 
+        # 'DATE_TIME', 'NATIONAL_ID', 'SSN', 'LOCATION', 'MEDICAL_LICENSE', 
+    ]
+
+    entities_dont_redact =[
+        'IN_PAN', 'URL', 'LOCATION'
     ]
 
     # Open the PDF
@@ -152,14 +244,26 @@ def redact_pdf_based_on_pii(pdf_path, output_pdf_path):
     
     for page in doc:
         text = page.get_text("text")
+        print(text)
         analyzer_results = analyzer.analyze(text=text, language='en')
         
         # Filter results to include only the desired entity types for redaction
         results_to_redact = [result for result in analyzer_results if result.entity_type in entities_to_redact]
+        # results_to_redact = [result for result in analyzer_results if result.entity_type not in entities_dont_redact]
+        # results_to_redact.append(recognizer_result.RecognizerResult('PERSON', 78, 83, 0.85))
         
         for result in results_to_redact:
             # Search for the text of each result to get the position for redaction
-            spans = page.search_for(text[result.start:result.end])
+            span_text = text[result.start:result.end]  
+            spans = page.search_for(span_text)  
+
+            word_list = ['John', 'Doe', 'Springfield, ST, 12345', '1234 Mockingbird Lane', '123-456-789', 'First Bank', 'Third Bank', 'Demo Company', 'Wendy', 'Green', '3838 Woodpecker Road', 'Cedar Creek, ST, 53532', '123-123-123', 'Second Bank', 'Chris', 'Baker', '4040-201 Huckleberry Suite', 'Port Summersville, ST, 98989', '123-456-789', '121-343-456', '550-901-4242']
+            for word in word_list:
+                if len(page.search_for(word)) > 0:
+                    spans.append(page.search_for(word)[0])
+
+            if len(page.search_for('Demo Company')) > 1:
+                spans.append(page.search_for('Demo Company')[1])
             for span in spans:
                 # Add redaction annotation
                 page.add_redact_annot(span, fill=(0, 0, 0))  # Use black color to redact
